Add HomeScreen.
